ai_researcher/requirements.txt
# PDF Processing & Metadata
pymupdf
pymupdf4llm
marker-pdf
weasyprint # <-- Add HTML to PDF conversion
python-docx # <-- Add DOCX creation

# Embeddings & Vector Store
FlagEmbedding
chromadb
torch
torchvision
torchaudio
numpy

# LLM Interaction
openai
python-dotenv
requests
httpx # <-- Add async HTTP client

# Agentic Layer & Schemas
pydantic

# UI
streamlit
gradio

# Utilities
tqdm
markdown
aiofiles # <-- Add async file I/O library
tavily-python
newspaper3k # <-- Add web content extraction library
lxml[html_clean] # <-- Add lxml with cleaning extras for newspaper3k
lxml_html_clean